Removed DocumentRouteMixin, moved logic to AuthenticatedRouteMixin
